fix: eval prompt route refactored
